Fix issue in -[AKAKeyboardActivationSequence deactivate] that could result in the sequence's actions all becoming ineffective until the keyboard is dismissed using a mechanism outside of the AKAKeyboardActivationSequenceAccessoryView. Fix issue in -[AKAKeyboardActivationSequence activateNext/Previous] that could result in the input accessory view's next/previous arrows not accurately indicating the correct state.

+++ AKABeacon/AKABeacon/Classes/AKAOperationObserver.h
@@ -30,8 +30,8 @@
  * @param workloadDifference a value specifying the amount by which the workload has changed.
  */
 - (void)                operation:(AKAOperation*)operation
-                didUpdateProgress:(CGFloat)progressDifference
-                         workload:(CGFloat)workloadDifference;
+                didUpdateProgress:(float)progressDifference
+                         workload:(float)workloadDifference;
 
 @optional
 - (void)                operation:(AKAOperation*)operation
